+++ packages/tests/test/singleton_test.dart
@@ -1,13 +1,13 @@
 import 'package:test/test.dart';
 import 'package:shsp/src/impl/utility/message_callback_map_singleton.dart';
 import 'package:shsp/src/impl/utility/shsp_socket_info_singleton.dart';
-import 'package:shsp/src/impl/shsp_instance/shsp_instance_handler_singleton.dart';
+import 'package:shsp/src/impl/instance/shsp_instance_handler_singleton.dart';
 
 void main() {
   group('Singletons', () {
     setUp(() {
-      MessageCallbackMapSingleton.destroy();
-      ShspSocketInfoSingleton.destroy();
+      MessageCallbackMapSingleton().destroy();
+      ShspSocketInfoSingleton().destroy();
       ShspInstanceHandlerSingleton.destroy();
     });